plutus-tx: split up the prelude a bit
