+++ plutus-tx/src/Language/PlutusTx/Eq.hs
@@ -1,0 +1,61 @@
+{-# OPTIONS_GHC -fno-omit-interface-pragmas #-}
+module Language.PlutusTx.Eq (Eq(..)) where
+
+import           Language.PlutusTx.Bool
+import qualified Language.PlutusTx.Builtins as Builtins
+import           Prelude                    hiding (Eq (..), not, (&&))
+
+{-# ANN module ("HLint: ignore"::String) #-}
+
+-- Copied from the GHC definition
+-- | The 'Eq' class defines equality ('==') and inequality ('/=').
+--
+-- Minimal complete definition: either '==' or '/='.
+--
+class Eq a where
+    (==), (/=)           :: a -> a -> Bool
+
+    {-# INLINABLE (/=) #-}
+    x /= y               = not (x == y)
+    {-# INLINABLE (==) #-}
+    x == y               = not (x /= y)
+
+instance Eq Integer where
+    {-# INLINABLE (==) #-}
+    (==) = Builtins.equalsInteger
+
+instance Eq Builtins.ByteString where
+    {-# INLINABLE (==) #-}
+    (==) = Builtins.equalsByteString
+
+instance Eq a => Eq [a] where
+    {-# INLINABLE (==) #-}
+    [] == [] = True
+    (x:xs) == (y:ys) = x == y && xs == ys
+    _ == _ = False
+
+instance Eq Bool where
+    {-# INLINABLE (==) #-}
+    True == True = True
+    False == False = True
+    _ == _ = False
+
+instance Eq a => Eq (Maybe a) where
+    {-# INLINABLE (==) #-}
+    (Just a1) == (Just a2) = a1 == a2
+    Nothing == Nothing = True
+    _ == _ = False
+
+instance (Eq a, Eq b) => Eq (Either a b) where
+    {-# INLINABLE (==) #-}
+    (Left a1) == (Left a2) = a1 == a2
+    (Right b1) == (Right b2) = b1 == b2
+    _ == _ = False
+
+instance Eq () where
+    {-# INLINABLE (==) #-}
+    _ == _ = True
+
+instance (Eq a, Eq b) => Eq (a, b) where
+    {-# INLINABLE (==) #-}
+    (a, b) == (a', b') = a == a' && b == b'
