+++ plutus-tx/src/Language/PlutusTx/Prelude.hs
@@ -4,13 +4,9 @@
 module Language.PlutusTx.Prelude (
     -- $prelude
     -- * Classes
-    Eq (..),
-    Ord (..),
-    Functor (..),
-    (<$>),
-    (<$),
-    -- * Functions
-    const,
+    module Eq,
+    module Ord,
+    module Functor,
     -- * String and tracing functions
     toPlutusString,
     trace,
@@ -21,10 +17,8 @@ module Language.PlutusTx.Prelude (
     -- * Error
     error,
     check,
-    -- * Boolean operators
-    (&&),
-    (||),
-    not,
+    -- * Booleans
+    module Bool,
     -- * Int operators
     plus,
     minus,
@@ -35,20 +29,9 @@ module Language.PlutusTx.Prelude (
     fst,
     snd,
     -- * Maybe
-    isJust,
-    isNothing,
-    maybe,
-    mapMaybe,
+    module Maybe,
     -- * Lists
-    null,
-    map,
-    foldr,
-    foldl,
-    length,
-    all,
-    any,
-    (++),
-    filter,
+    module List,
     -- * ByteStrings
     ByteString,
     takeByteString,
@@ -62,10 +45,16 @@ module Language.PlutusTx.Prelude (
     module Prelude
     ) where
 
+import           Language.PlutusTx.Bool     as Bool
 import           Language.PlutusTx.Builtins (ByteString, concatenate, dropByteString, emptyByteString, equalsByteString,
                                              greaterThanByteString, lessThanByteString, sha2_256, sha3_256,
                                              takeByteString, verifySignature)
 import qualified Language.PlutusTx.Builtins as Builtins
+import           Language.PlutusTx.Eq       as Eq
+import           Language.PlutusTx.Functor  as Functor
+import           Language.PlutusTx.List     as List
+import           Language.PlutusTx.Maybe    as Maybe
+import           Language.PlutusTx.Ord      as Ord
 import           Prelude                    as Prelude hiding (Eq (..), Functor (..), Ord (..), all, any, const, error,
                                                         filter, foldl, foldr, fst, length, map, max, maybe, min, not,
                                                         null, snd, (&&), (++), (<$>), (||))
@@ -85,201 +74,6 @@ import           Prelude                    as Prelude hiding (Eq (..), Functor 
 -- $setup
 -- >>> :set -XNoImplicitPrelude
 -- >>> import Language.PlutusTx.Prelude
-
--- Copied from the GHC definition
--- | The 'Eq' class defines equality ('==') and inequality ('/=').
---
--- Minimal complete definition: either '==' or '/='.
---
-class Eq a where
-    (==), (/=)           :: a -> a -> Bool
-
-    {-# INLINABLE (/=) #-}
-    x /= y               = not (x == y)
-    {-# INLINABLE (==) #-}
-    x == y               = not (x /= y)
-
-instance Eq Integer where
-    {-# INLINABLE (==) #-}
-    (==) = Builtins.equalsInteger
-
-instance Eq ByteString where
-    {-# INLINABLE (==) #-}
-    (==) = Builtins.equalsByteString
-
-instance Eq a => Eq [a] where
-    {-# INLINABLE (==) #-}
-    [] == [] = True
-    (x:xs) == (y:ys) = x == y && xs == ys
-    _ == _ = False
-
-instance Eq Bool where
-    {-# INLINABLE (==) #-}
-    True == True = True
-    False == False = True
-    _ == _ = False
-
-instance Eq a => Eq (Maybe a) where
-    {-# INLINABLE (==) #-}
-    (Just a1) == (Just a2) = a1 == a2
-    Nothing == Nothing = True
-    _ == _ = False
-
-instance (Eq a, Eq b) => Eq (Either a b) where
-    {-# INLINABLE (==) #-}
-    (Left a1) == (Left a2) = a1 == a2
-    (Right b1) == (Right b2) = b1 == b2
-    _ == _ = False
-
-instance Eq () where
-    {-# INLINABLE (==) #-}
-    _ == _ = True
-
-instance (Eq a, Eq b) => Eq (a, b) where
-    {-# INLINABLE (==) #-}
-    (a, b) == (a', b') = a == a' && b == b'
-
--- Copied from the GHC definition
--- | The 'Ord' class is used for totally ordered datatypes.
---
--- Minimal complete definition: either 'compare' or '<='.
--- Using 'compare' can be more efficient for complex types.
---
-class Eq a => Ord a where
-    compare              :: a -> a -> Ordering
-    (<), (<=), (>), (>=) :: a -> a -> Bool
-    max, min             :: a -> a -> a
-
-    {-# INLINABLE compare #-}
-    compare x y = if x == y then EQ
-                  -- NB: must be '<=' not '<' to validate the
-                  -- above claim about the minimal things that
-                  -- can be defined for an instance of Ord:
-                  else if x <= y then LT
-                  else GT
-
-    {-# INLINABLE (<) #-}
-    x <  y = case compare x y of { LT -> True;  _ -> False }
-    {-# INLINABLE (<=) #-}
-    x <= y = case compare x y of { GT -> False; _ -> True }
-    {-# INLINABLE (>) #-}
-    x >  y = case compare x y of { GT -> True;  _ -> False }
-    {-# INLINABLE (>=) #-}
-    x >= y = case compare x y of { LT -> False; _ -> True }
-
-    -- These two default methods use '<=' rather than 'compare'
-    -- because the latter is often more expensive
-    {-# INLINABLE max #-}
-    max x y = if x <= y then y else x
-    {-# INLINABLE min #-}
-    min x y = if x <= y then x else y
-
-instance Ord Integer where
-    {-# INLINABLE (<) #-}
-    (<) = Builtins.lessThanInteger
-    {-# INLINABLE (<=) #-}
-    (<=) = Builtins.lessThanEqInteger
-    {-# INLINABLE (>) #-}
-    (>) = Builtins.greaterThanInteger
-    {-# INLINABLE (>=) #-}
-    (>=) = Builtins.greaterThanEqInteger
-
-instance Ord ByteString where
-    {-# INLINABLE compare #-}
-    compare l r = if Builtins.lessThanByteString l r then LT else if Builtins.equalsByteString l r then EQ else GT
-
-instance Ord a => Ord [a] where
-    {-# INLINABLE compare #-}
-    compare []     []     = EQ
-    compare []     (_:_)  = LT
-    compare (_:_)  []     = GT
-    compare (x:xs) (y:ys) = case compare x y of
-                                EQ    -> compare xs ys
-                                other -> other
-
-instance Ord Bool where
-    {-# INLINABLE compare #-}
-    compare b1 b2 = case b1 of
-        False -> case b2 of
-            False -> EQ
-            True  -> LT
-        True -> case b2 of
-            False -> GT
-            True  -> EQ
-
-instance Ord a => Ord (Maybe a) where
-    {-# INLINABLE compare #-}
-    compare (Just a1) (Just a2) = compare a1 a2
-    compare Nothing (Just _)    = LT
-    compare (Just _) Nothing    = GT
-    compare Nothing Nothing     = EQ
-
-instance (Ord a, Ord b) => Ord (Either a b) where
-    {-# INLINABLE compare #-}
-    compare (Left a1) (Left a2)   = compare a1 a2
-    compare (Left _) (Right _)    = LT
-    compare (Right _) (Left _)    = GT
-    compare (Right b1) (Right b2) = compare b1 b2
-
-instance Ord () where
-    {-# INLINABLE compare #-}
-    compare _ _ = EQ
-
-instance (Ord a, Ord b) => Ord (a, b) where
-    {-# INLINABLE compare #-}
-    compare (a, b) (a', b') = case compare a a' of
-        LT -> LT
-        GT -> GT
-        EQ -> compare b b'
-
-{- | The 'Functor' class is used for types that can be mapped over.
-Instances of 'Functor' should satisfy the following laws:
-
-> fmap id  ==  id
-> fmap (f . g)  ==  fmap f . fmap g
--}
-class Functor f where
-    fmap :: (a -> b) -> f a -> f b
-
-    -- <$ deliberately omitted, to make this a one-method class which has a
-    -- simpler representation
-
-infixl 4 <$>
--- | An infix synonym for 'fmap'.
-{-# INLINABLE (<$>) #-}
-(<$>) :: Functor f => (a -> b) -> f a -> f b
-(<$>) f fa = fmap f fa
-
-infixl 4 <$
-{-# INLINABLE (<$) #-}
--- | Replace all locations in the input with the same value.
-(<$) :: Functor f => a -> f b -> f a
-(<$) a fb = fmap (const a) fb
-
-instance Functor [] where
-    {-# INLINABLE fmap #-}
-    fmap f l = case l of
-            []   -> []
-            x:xs -> f x : fmap f xs
-
-instance Functor Maybe where
-    {-# INLINABLE fmap #-}
-    fmap f (Just a) = Just (f a)
-    fmap _ Nothing  = Nothing
-
-instance Functor (Either c) where
-    {-# INLINABLE fmap #-}
-    fmap f (Right a) = Right (f a)
-    fmap _ (Left c)  = Left c
-
-instance Functor ((,) c) where
-    {-# INLINABLE fmap #-}
-    fmap f (c, a) = (c, f a)
-
-{-# INLINABLE const #-}
--- | Plutus Tx version of 'Prelude.const'.
-const                   :: a -> b -> a
-const x _               =  x
 
 {-# INLINABLE error #-}
 -- | Terminate the evaluation of the script with an error message.
@@ -326,35 +120,6 @@ traceIfFalseH str a = if a then True else traceH str False
 traceIfTrueH :: String -> Bool -> Bool
 traceIfTrueH str a = if a then traceH str True else False
 
-{-# INLINABLE (&&) #-}
--- | Logical AND
---
---   >>> True && False
---   False
---
-infixr 3 &&
-(&&) :: Bool -> Bool -> Bool
-(&&) l r = if l then r else False
-
-{-# INLINABLE (||) #-}
--- | Logical OR
---
---   >>> True || False
---   True
---
-infixr 2 ||
-(||) :: Bool -> Bool -> Bool
-(||) l r = if l then True else r
-
-{-# INLINABLE not #-}
--- | Logical negation
---
---   >>> not True
---   False
---
-not :: Bool -> Bool
-not a = if a then False else True
-
 {-# INLINABLE plus #-}
 -- | Addition
 --
@@ -399,142 +164,6 @@ divide = Builtins.divideInteger
 --
 remainder :: Integer -> Integer -> Integer
 remainder = Builtins.remainderInteger
-
-{-# INLINABLE isJust #-}
--- | Check if a 'Maybe' @a@ is @Just a@
---
---   >>> isJust Nothing
---   False
---   >>> isJust (Just "plutus")
---   True
---
-isJust :: Maybe a -> Bool
-isJust m = case m of { Just _ -> True; _ -> False; }
-
-{-# INLINABLE isNothing #-}
--- | Check if a 'Maybe' @a@ is @Nothing@
---
---   >>> isNothing Nothing
---   True
---   >>> isNothing (Just "plutus")
---   False
---
-isNothing :: Maybe a -> Bool
-isNothing m = case m of { Just _ -> False; _ -> True; }
-
-{-# INLINABLE maybe #-}
--- | PlutusTx version of 'Prelude.maybe'.
---
---   >>> maybe "platypus" (\s -> s) (Just "plutus")
---   "plutus"
---   >>> maybe "platypus" (\s -> s) Nothing
---   "platypus"
---
-maybe :: b -> (a -> b) -> Maybe a -> b
-maybe b f m = case m of
-    Nothing -> b
-    Just a  -> f a
-
-{-# INLINABLE null #-}
--- | PlutusTx version of 'Data.List.null'.
---
---   >>> null [1]
---   False
---   >>> null []
---   True
---
-null :: [a] -> Bool
-null l = case l of
-    [] -> True
-    _  -> False
-
-{-# INLINABLE map #-}
--- | PlutusTx version of 'Data.List.map'.
---
---   >>> map (\i -> i + 1) [1, 2, 3]
---   [2,3,4]
---
-map :: (a -> b) -> [a] -> [b]
-map f l = case l of
-    []   -> []
-    x:xs -> f x : map f xs
-
-{-# INLINABLE foldr #-}
--- | PlutusTx version of 'Data.List.foldr'.
---
---   >>> foldr (\i s -> s + i) 0 [1, 2, 3, 4]
---   10
---
-foldr :: (a -> b -> b) -> b -> [a] -> b
-foldr f acc l = case l of
-    []   -> acc
-    x:xs -> f x (foldr f acc xs)
-
-{-# INLINABLE foldl #-}
--- | PlutusTx version of 'Data.List.foldl'.
---
---   >>> foldl (\s i -> s + i) 0 [1, 2, 3, 4]
---   10
---
-foldl :: (b -> a -> b) -> b -> [a] -> b
-foldl f acc l = case l of
-    []   -> acc
-    x:xs -> foldl f (f acc x) xs
-
-{-# INLINABLE length #-}
--- | 'length' @xs@ is the number of elements in @xs@.
---
---   >>> length [1, 2, 3, 4]
---   4
---
-length :: [a] -> Integer
--- eta-expanded to avoid the value restriction
-length as = foldr (\_ acc -> plus acc 1) 0 as
-
-{-# INLINABLE all #-}
--- | PlutusTx version of 'Data.List.all'.
---
---   >>> all (\i -> i > 5) [6, 8, 12]
---   True
---
-all :: (a -> Bool) -> [a] -> Bool
-all p = foldr (\a acc -> acc && p a) True
-
-{-# INLINABLE any #-}
--- | PlutusTx version of 'Data.List.any'.
---
---   >>> any (\i -> i > 5) [6, 2, 1]
---   True
---
-any :: (a -> Bool) -> [a] -> Bool
-any p = foldr (\a acc -> acc || p a) False
-
-{-# INLINABLE (++) #-}
--- | PlutusTx version of 'Data.List.(++)'.
---
---   >>> [0, 1, 2] ++ [1, 2, 3, 4]
---   [0,1,2,1,2,3,4]
---
-(++) :: [a] -> [a] -> [a]
-(++) l r = foldr (:) r l
-
-{-# INLINABLE filter #-}
--- | PlutusTx version of 'Data.List.filter'.
---
---   >>> filter (> 1) [1, 2, 3, 4]
---   [2,3,4]
---
-filter :: (a -> Bool) -> [a] -> [a]
-filter p = foldr (\e xs -> if p e then e:xs else xs) []
-
-{-# INLINABLE mapMaybe #-}
--- | PlutusTx version of 'Data.Maybe.mapMaybe'.
---
---   >>> mapMaybe (\i -> if i == 2 then Just '2' else Nothing) [1, 2, 3, 4]
---   "2"
---
-mapMaybe :: (a -> Maybe b) -> [a] -> [b]
-mapMaybe p = foldr (\e xs -> case p e of { Just e' -> e':xs; Nothing -> xs}) []
 
 {-# INLINABLE fst #-}
 -- | PlutusTx version of 'Data.Tuple.fst'
